Removed deprecated tests-as-filters (#14)
* Removed deprecated tests-as-filters

There were a few instances of `result | failed` (or similar) which I've
removed. I also factored out an extremely long line in the "Download
get-pip.py" line to a variable and added a 'pip_executable' variable
(even after pip was installed, a misconfigured PATH would make 'pip
--version' still fail).

* Bumped min_ansible_version from 1.8 to 2.5

tests-as-filters were deprecated with Ansible 2.5 and will be removed in
Ansible 2.9.